Split SearchView and VideoDetailView into component files
Extract SearchFilterSheet, SearchTextFieldReturnKeyEnabler, and
SearchResultRow from SearchView.swift (634→263 lines). Extract
RelatedVideoComponents from VideoDetailView.swift (982→736 lines).
