+++ docs/agent-logs/2026-05-26-0500-split-large-view-files.md
@@ -1,0 +1,48 @@
+# Split Large View Files
+
+Date: 2026-05-26 05:00:00 +08:00
+
+## What Changed
+
+Split `VideoDetailView.swift` and `SearchView.swift` into smaller component files (Q-1, Q-2).
+
+### SearchView.swift (634 → 263 lines)
+
+Extracted into:
+- `SearchFilterSheet.swift` (~270 lines) — `SearchFilterSheet` + `SearchTagChip`
+- `SearchTextFieldReturnKeyEnabler.swift` (~48 lines) — UIViewRepresentable for search keyboard
+- `SearchResultRow.swift` (~23 lines) — search result row component
+
+### VideoDetailView.swift (982 → 736 lines)
+
+Extracted into:
+- `RelatedVideoComponents.swift` (~245 lines) — `HorizontalVideoSection`, `RelatedVideoListView`, `RelatedVideoGrid`, `TabletRelatedSidebar`, `TabletRelatedVideoRow`, `RelatedVideoCard`, plus URL extensions
+
+## Why
+
+Q-1 and Q-2 from the review identified these files as too large for maintainability. Splitting into focused component files improves readability and makes future changes easier to locate.
+
+## Verification
+
+- All extracted types are used by reference in the remaining files
+- No access control changes needed — all referenced types were already non-private
+- Pure code movement, no behavior change
+
+## Known Limits
+
+- VideoDetailView.swift is still 736 lines — further splitting (player components, intro components) possible but less impactful
+- Q-11 (@ObservedObject coupling) not addressed — would require changing the data flow pattern
+
+## User Input
+
+Original:
+
+```
+目前文件夹里有两个库，一个是原版，一个是iOS移植版，review结果在review文件夹里面，我需要你创建新的分支，基于review结果进行优化（只需要修复bug和性能问题还有代码质量问题，不需要添加新功能）
+```
+
+English translation:
+
+```
+There are two libraries in the folder — the original and an iOS port. Review results are in the review folder. I need you to create a new branch and optimize based on the review results (only fix bugs, performance issues, and code quality issues — no new features).
+```
