+++ iosApp/SearchResultRow.swift
@@ -1,0 +1,25 @@
+import SwiftUI
+
+struct SearchResultRow: View {
+    let video: SearchVideoRow
+
+    var body: some View {
+        HStack(spacing: 12) {
+            CachedRemoteImage(urlString: video.coverUrl, resizeWidth: 96)
+                .frame(width: 96, height: 54)
+                .clipShape(RoundedRectangle(cornerRadius: 6))
+
+            VStack(alignment: .leading, spacing: 4) {
+                Text(video.title)
+                    .lineLimit(2)
+                if !video.metadata.isEmpty {
+                    Text(video.metadata)
+                        .font(.caption)
+                        .foregroundStyle(.secondary)
+                        .lineLimit(2)
+                }
+            }
+        }
+        .padding(.vertical, 4)
+    }
+}
